cambie auxiliares. variable y funciones y pase a fech search and detalles
