+++ src/auxiliares/funciones.js
@@ -1,0 +1,4 @@
+// Definir URL entre categoria, tipoy week
+export const definirURLGeneral = (categoria, tipo, week) => {
+	week ? `${categoria}/${tipo}/${week}` : `${tipo}/${categoria}`;
+};
