+++ src/auxiliares/variables.js
@@ -1,0 +1,4 @@
+export const urlBase = `https://api.themoviedb.org/3/`;
+export const apiKey = `api_key=92b7c9e2808de339886a0b75ca3aa28e`;
+export const paginas = "&page=";
+export const lenguaje = "&language=";
