--- src/hooks/UseFecthPaginado.js
@@ -1,4 +1,0 @@
-const paginado = () => {
-	return <div>hola</div>;
-};
-export default paginado;
