+++ src/hooks/UseFetchGeneral.js
@@ -1,11 +1,6 @@
 import { useState, useEffect } from "react";
-import {
-	urlBase,
-	apiKey,
-	lenguaje,
-	paginas,
-	definirURL,
-} from "../auxiliares/auxiliares";
+import { urlBase, apiKey, lenguaje, paginas } from "../auxiliares/variables";
+import { definirURLGeneral } from ".../auxiliares/variables";
 
 //https://api.themoviedb.org/3/${categorias}/${tipo}${week}?api_key=92b7c9e2808de339886a0b75ca3aa28e&language=es-AR&page=1`
 
@@ -14,7 +9,7 @@ const UseFetchGeneral = (categoria, tipo, page) => {
 	const [isLoading, setIsLoading] = useState(false);
 	const [totalPages, setTotalPages] = useState(1);
 
-	const url = `${urlBase}${definirURL(
+	const url = `${urlBase}${definirURLGeneral(
 		categoria,
 		tipo,
 		categoria === "trending" ? "week" : ""
